Move dropdown popup position test to JVM
