+++ composeunstyled-dropdown-menu/src/commonTest/kotlin/com/composeunstyled/DropdownMenuPanelTest.kt
@@ -21,11 +21,8 @@
  */
 package com.composeunstyled
 
-import androidx.compose.animation.EnterTransition
-import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.core.MutableTransitionState
 import androidx.compose.foundation.focusable
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
@@ -36,8 +33,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.InputMode
 import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.layout.onPlaced
-import androidx.compose.ui.layout.positionInWindow
 import androidx.compose.ui.platform.LocalInputModeManager
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.assertIsFocused
@@ -48,51 +43,11 @@ import androidx.compose.ui.test.performKeyInput
 import androidx.compose.ui.test.pressKey
 import androidx.compose.ui.test.requestFocus
 import androidx.compose.ui.test.runComposeUiTest
-import androidx.compose.ui.unit.dp
 import kotlin.test.Test
-import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 class DropdownMenuPanelTest {
-  @Test
-  fun expandedMenuPlacesPanelAtAnchorPositionOnFirstFrame() = runComposeUiTest {
-    mainClock.autoAdvance = false
-    val panelPositions = mutableListOf<Int>()
-
-    setContent {
-      UnstyledDropdownMenu(
-        expanded = true,
-        onExpandedChange = {},
-        alignment = AnchorAlignment.End,
-        panel = {
-          DropdownMenuPanel(
-            enter = EnterTransition.None,
-            exit = ExitTransition.None,
-          ) {
-            BasicText(
-              text = "Menu content",
-              modifier = Modifier
-                .size(width = 60.dp, height = 40.dp)
-                .onPlaced { coordinates ->
-                  panelPositions += coordinates.positionInWindow().x.toInt()
-                },
-            )
-          }
-        },
-        anchor = {
-          BasicText(
-            text = "Anchor",
-            modifier = Modifier.size(width = 80.dp, height = 40.dp),
-          )
-        },
-      )
-    }
-    mainClock.advanceTimeByFrame()
-
-    assertEquals(listOf(20), panelPositions)
-  }
-
   @Test
   fun panelWithoutMenuStateIsHidden() = runComposeUiTest {
     setContent {
