+++ composeunstyled-dropdown-menu/src/jvmTest/kotlin/com/composeunstyled/DropdownMenuPanelJvmTest.kt
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) 2026 Composable Horizons
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.composeunstyled
+
+import androidx.compose.animation.EnterTransition
+import androidx.compose.animation.ExitTransition
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.onPlaced
+import androidx.compose.ui.layout.positionInWindow
+import androidx.compose.ui.test.runComposeUiTest
+import androidx.compose.ui.unit.dp
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class DropdownMenuPanelJvmTest {
+  @Test
+  fun expandedMenuPlacesPanelAtAnchorPositionOnFirstFrame() = runComposeUiTest {
+    mainClock.autoAdvance = false
+    val panelPositions = mutableListOf<Int>()
+
+    setContent {
+      UnstyledDropdownMenu(
+        expanded = true,
+        onExpandedChange = {},
+        alignment = AnchorAlignment.End,
+        panel = {
+          DropdownMenuPanel(
+            enter = EnterTransition.None,
+            exit = ExitTransition.None,
+          ) {
+            BasicText(
+              text = "Menu content",
+              modifier = Modifier
+                .size(width = 60.dp, height = 40.dp)
+                .onPlaced { coordinates ->
+                  panelPositions += coordinates.positionInWindow().x.toInt()
+                },
+            )
+          }
+        },
+        anchor = {
+          BasicText(
+            text = "Anchor",
+            modifier = Modifier.size(width = 80.dp, height = 40.dp),
+          )
+        },
+      )
+    }
+    mainClock.advanceTimeByFrame()
+
+    assertEquals(listOf(20), panelPositions)
+  }
+}
